+++ tests/integration/fibonacci.cpp
@@ -2,10 +2,10 @@
 #include <iostream>
 
 extern "C" {
-double fib(double);
+float fib(float);
 }
 
 int main() {
-  double fib_10 = fib(10.);
+  float fib_10 = fib(10.);
   std::cout << "fib(10.): " << fib_10 << std::endl;
 }
